Individual Character Movement
characters can move on their own when selected
