+++ Assets/Objects/Characters/Prefabs/earth_mage.prefab
@@ -1480,7 +1480,7 @@ MonoBehaviour:
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1000011572547674}
-  m_Enabled: 1
+  m_Enabled: 0
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 785b86ae9fb119940bf5392d14556909, type: 3}
   m_Name:
