+++ Assets/Scripts/Player.cs
@@ -9,10 +9,16 @@ using System;
 /// </summary>
 public class Player : MonoBehaviour
 {
-
     public GameObject[] CharacterPrefabs = new GameObject[4];
 
     private GameObject[] characters = new GameObject[4];
+    private GameObject selectedHero;
+
+    public GameObject SelectedHero
+    {
+        get { return selectedHero;  }
+        set { this.selectedHero = value; }
+    }
 
     public GameObject[] Characters
     {
@@ -29,6 +35,31 @@ public class Player : MonoBehaviour
         }
     }
 
+    void Update()
+    {
+        PlayerInput.Type playerInput = PlayerInput.getPlayerInput();
+
+        // detect clicks only for the local player otherwise one player will control all characters
+        if (playerInput == PlayerInput.Type.MOVE && this.selectedHero != null)
+        {
+            FindAndSetTargetPosition();
+        }
+    }
+
+    /// <summary>
+    /// Sets the target position we will travel too.
+    /// </summary>
+    private void FindAndSetTargetPosition()
+    {
+        Plane plane = new Plane(Vector3.up, transform.position);
+        Ray ray = Camera.main.ScreenPointToRay(Input.mousePosition);
+        float point = 0f;
+
+        if (plane.Raycast(ray, out point))
+        {
+            this.selectedHero.GetComponent<CharacterControl>().SetTargetPosition(ray, point);
+        }
+    }
 
     Vector3[] getSpawnPoints()
     {
