+++ Assets/Scripts/PlayerInput.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 41c8fd0c086d6ed428e901db3186687a
+timeCreated: 1476161233
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
